refactor(umi): extract read-name UMI normalization into shared module
Move normalize_read_name_umi (and its REVCOMP_PREFIX/UMI_DELIMITER
constants and the is_valid_umi_char helper) out of the extract command
into a new src/lib/umi/read_name.rs as a crate-visible free function.
The logic is unchanged and remains byte-for-byte faithful to fgbio's
Umis.extractUmisFromReadName; extract now calls the shared function.

This is a pure move with no behavior change, preparing for a second
caller (the forthcoming copy-umi command) to reuse the same
normalization. Adds focused unit tests for the normalizer in its new
home, where it previously had only indirect coverage through extract.
