+++ src/lib/commands/extract.rs
@@ -37,7 +37,6 @@ use bstr::{BString, ByteSlice};
 use clap::Parser;
 use fgumi_bam_io::ProgressTracker;
 use fgumi_bam_io::{RawBamWriter, create_raw_bam_writer, open_output_writer};
-use fgumi_dna::dna::reverse_complement;
 use fgumi_raw_bam::UnmappedSamBuilder;
 use fgumi_raw_bam::fields::flags;
 use log::{debug, info};
@@ -71,14 +70,6 @@ const QUALITY_DETECTION_SAMPLE_SIZE: usize = 400;
 /// compressed BGZF blocks. Sized well above `BGZF_MAX_BLOCK_SIZE` so a block
 /// lands in the buffer instead of passing straight through it.
 const PIPELINE_OUTPUT_BUF_CAPACITY: usize = 256 * 1024;
-
-/// Prefix marking a read-name UMI segment that must be reverse-complemented,
-/// matching fgbio `Umis.RevcompPrefix` (`Umis.scala:33`).
-const REVCOMP_PREFIX: u8 = b'r';
-
-/// Delimiter separating dual UMIs in a read name, translated to `-` on extraction,
-/// matching fgbio's `umiDelimiter` default (`Umis.scala:88`).
-const UMI_DELIMITER: u8 = b'+';
 
 /// Pre-serialized batch of BAM records for the pipeline output type.
 ///
@@ -929,7 +920,7 @@ impl Extract {
     ///
     /// Names with fewer than 8 fields are treated as not containing a UMI and
     /// produce `None`. The extracted UMI is then normalized via
-    /// [`normalize_read_name_umi`](Self::normalize_read_name_umi) to match fgbio's
+    /// [`normalize_read_name_umi`](crate::umi::read_name::normalize_read_name_umi) to match fgbio's
     /// strict `Umis.extractUmisFromReadName` (reverse-complement `r`-prefixed
     /// segments, `+`→`-`, upper-case).
     ///
@@ -966,83 +957,13 @@ impl Extract {
             && let Some(last) = parts.last()
             && !last.is_empty()
         {
-            let umi = Self::normalize_read_name_umi(last).with_context(|| {
+            let umi = crate::umi::read_name::normalize_read_name_umi(last).with_context(|| {
                 format!("extracting UMI from read name '{}'", String::from_utf8_lossy(name_part))
             })?;
             return Ok((name_part.to_vec(), Some(umi)));
         }
 
         Ok((name_part.to_vec(), None))
-    }
-
-    /// Normalizes a UMI extracted from the last field of a read name, matching
-    /// fgbio's `Umis.extractUmisFromReadName` (strict mode, as called by
-    /// `FastqToBam`; `Umis.scala:85-126`).
-    ///
-    /// With `reverseComplementPrefixedUmis` defaulting to true, fgbio:
-    /// 1. reverse-complements any `r`-prefixed segment (EXT-04); when the UMI is
-    ///    `+`-delimited, only the `r`-prefixed segments are reverse-complemented,
-    /// 2. translates the `+` dual-UMI delimiter to `-`,
-    /// 3. upper-cases the result (EXT-03), and
-    /// 4. throws if any character is outside `ACGTN-` (EXT-01, strict mode).
-    ///
-    /// Reverse-complementing reuses [`fgumi_dna::dna::reverse_complement`], which
-    /// complements `ACGT`, maps RNA `U`→`A`, and preserves `N` — matching fgbio's
-    /// `Sequences.complement` over the bases a read-name UMI may contain. Any other
-    /// character is passed through and then rejected by the `ACGTN-` validation
-    /// below, mirroring fgbio's strict-mode rejection.
-    ///
-    /// # Errors
-    /// Returns an error if the normalized UMI contains a character outside `ACGTN-`,
-    /// mirroring fgbio's strict-mode `IllegalArgumentException`.
-    fn normalize_read_name_umi(raw: &[u8]) -> Result<Vec<u8>> {
-        // fgbio keys the transform on whether an 'r' appears anywhere and whether a
-        // '+' delimiter appears at an index > 0 (a leading '+' is not a delimiter).
-        let has_revcomp_prefix = raw.contains(&REVCOMP_PREFIX);
-        let has_delimiter = raw.iter().position(|&b| b == UMI_DELIMITER).is_some_and(|idx| idx > 0);
-
-        let transformed: Vec<u8> = match (has_revcomp_prefix, has_delimiter) {
-            // Reverse-complement each `r`-prefixed segment, join with '-'.
-            (true, true) => {
-                let mut out = Vec::with_capacity(raw.len());
-                for (i, segment) in raw.split(|&b| b == UMI_DELIMITER).enumerate() {
-                    if i > 0 {
-                        out.push(b'-');
-                    }
-                    match segment.split_first() {
-                        Some((&REVCOMP_PREFIX, rest)) => out.extend(reverse_complement(rest)),
-                        // `stripPrefix('r')` is a no-op when the segment is not
-                        // prefixed with 'r'; the segment is copied verbatim.
-                        _ => out.extend_from_slice(segment),
-                    }
-                }
-                out
-            }
-            // Single UMI: reverse-complement after stripping a leading 'r'.
-            (true, false) => {
-                let stripped = raw.strip_prefix(&[REVCOMP_PREFIX][..]).unwrap_or(raw);
-                reverse_complement(stripped)
-            }
-            // No revcomp: just translate the '+' delimiter to '-'.
-            (false, true) => {
-                raw.iter().map(|&b| if b == UMI_DELIMITER { b'-' } else { b }).collect()
-            }
-            (false, false) => raw.to_vec(),
-        };
-
-        // fgbio upper-cases the UMI after any reverse-complementing.
-        let umi: Vec<u8> = transformed.iter().map(u8::to_ascii_uppercase).collect();
-
-        // Strict mode: reject any character outside the SAM UMI alphabet.
-        if let Some(&bad) = umi.iter().find(|&&b| !is_valid_umi_char(b)) {
-            bail!(
-                "Invalid UMI '{}' extracted from read name (illegal character '{}')",
-                String::from_utf8_lossy(&umi),
-                bad as char,
-            );
-        }
-
-        Ok(umi)
     }
 
     /// Validates that all read names match across the read sets
@@ -1812,16 +1733,6 @@ impl Command for Extract {
         timer.log_completion(records_written);
         Ok(())
     }
-}
-
-/// Returns true if `b` is a valid SAM UMI character (`ACGTN-`), matching fgbio's
-/// `Umis.isValidUmiCharacter` (`Umis.scala:128-130`).
-///
-/// This is intentionally distinct from [`fgumi_umi::validate_umi`], which
-/// implements the more permissive `GroupReadsByUmi` counting rule (only rejects
-/// upper-case `N`); the read-name extraction path uses fgbio's strict alphabet.
-fn is_valid_umi_char(b: u8) -> bool {
-    matches!(b, b'A' | b'C' | b'G' | b'T' | b'N' | b'-')
 }
 
 /// Reject an assembled template that does not carry exactly one record per read
